+++ app/enums.py
@@ -11,3 +11,14 @@ class SubState(enum.Enum):
     subscribed = enum.auto()
     trial = enum.auto()
     outlaw = enum.auto()
+
+
+class PaymentStatus(enum.Enum):
+    """"The status of a Payment."""
+    manual = enum.auto()
+    creating = enum.auto()
+    waiting = enum.auto()
+    accepted = enum.auto()
+    refused = enum.auto()
+    cancelled = enum.auto()
+    error = enum.auto()